Builder is split in container groups, viewModel is cleaned up, and looks presentable. Currently trying to connect everything back.
