+++ css/build-maker.css
@@ -1,0 +1,94 @@
+/***************************************/
+/* BUILDER SECTION */
+/***************************************/
+.build-maker-container {
+	display: grid;
+
+	grid-template-columns: 1fr;
+	grid-template-rows: auto 1fr auto;
+	height: 100vh;
+
+	li {
+		list-style-type: none;
+	}
+}
+
+/* work-area */
+.build-maker {
+	display: grid;
+	grid-template-columns: auto 1fr auto;
+	grid-template-rows: auto 1fr auto;
+	gap: 10px;
+
+	position: relative;
+
+	z-index: 1;
+
+	background-color: var(--dark-color);
+
+	/* box */
+	.grid-container {
+		padding: 2rem;
+		border: 1px solid red;
+	}
+	.ship-info-container {
+		grid-row: 1 / 2;
+		grid-column: 1 / 2;
+	}
+
+	.fighter-container-container {
+		grid-column: 1 / 2;
+		grid-row: 2 / 3;
+	}
+	.builder-buttons-container {
+		grid-column: 1 / 2;
+		grid-row: 3 / -1;
+	}
+	/* center */
+	/* REMOVE */
+	.additional-info-container {
+		grid-row: 1 / 2;
+		grid-column: 2 / 3;
+	}
+	.hangar-container {
+		grid-row: 2 / -1;
+		grid-column: 2 / 3;
+	}
+
+	.stats-container {
+		/* border: 1px solid yellow; */
+		grid-column: 3 / -1;
+		grid-row: 1 / -1;
+	}
+
+	/* content in work area */
+	.box__top-left__CR-graph {
+		background-color: var(--accent-color);
+		width: 16rem;
+		height: 2rem;
+	}
+
+	/* Content Area Right */
+	.top-content-group ul {
+		margin-bottom: 1.6rem;
+	}
+
+	/*  Ship build Area */
+	.box__middle-center {
+		display: flex;
+		justify-content: center;
+		align-items: center;
+	}
+}
+/* BUILDER BUTTONS CONTAINER */
+.builder-buttons-container ul {
+	display: flex;
+	flex-direction: column;
+	gap: 1rem;
+}
+/* HANGAR CONTAINER */
+.hangar-container {
+	display: flex;
+	justify-content: center;
+	align-items: center;
+}
