+++ css/main.css
@@ -4,7 +4,10 @@
 @import "_icons.css";
 @import "_fonts.css";
 @import "_weapon-pop-up.css";
+
 @import "_build-maker.css";
+@import "./build-maker.css";
+@import "./navigation.css";
 
 @import "./Weapons/_weaponSprite.css";
 @import "./Weapons/_weaponBackgroundSprite.css";
