+++ script/components/Containers/AdditionalInfoContainerView.js
@@ -1,0 +1,11 @@
+import classNames from "../../helper/DomClassNames";
+
+class AdditionalInfoContainerView {
+	render() {
+		const localParent = `.${classNames.additionalInfoContainer}`;
+		const markup = "additional info";
+
+		return [markup, localParent];
+	}
+}
+export default new AdditionalInfoContainerView();
