+++ script/helper/DataSet.js
@@ -8,5 +8,6 @@ const DataSet = {
 	dataButtonValue: "data-button-value",
 	dataButtonName: "data-button-name",
 	dataHullModButtonType: "data-hullmod-button-type",
+	dataWeaponId: "data-weapon-id",
 };
 export default DataSet;
